Reorganise file structure
