+++ src/main/java/duke/parser/DeadlineParser.java
@@ -1,6 +1,7 @@
-package duke.models;
+package duke.parser;
 
 import duke.exceptions.DukeException;
+import duke.models.Deadline;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
