+++ src/test/java/duke/model/ParserTest.java
@@ -1,7 +1,7 @@
 package duke.model;
 
 import duke.handlers.AddTodoCommand;
-import duke.models.Parser;
+import duke.parser.Parser;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
